COSO

+++ script.js
@@ -1,17 +1,16 @@
-function compute()
-{
+function compute() {
     let principal = document.getElementById("principal").value;
     let rate = document.getElementById("rate").value;
     let years = document.getElementById("years").value;
-    
-    let interest = parseInt(principal) * parseInt(rate) *parseInt(years)/ 100;
-    let amount = parseInt( principal) + parseInt(interest);
-    let year = new Date().getFullYear() + parseInt(years);
-    
-    if (parseInt(principal) <= 0){
+
+    let interest = parseFloat(principal) * parseFloat(rate) * parseFloat(years) / 100;
+    let amount = parseFloat(principal) + parseFloat(interest);
+    let year = new Date().getFullYear() + parseFloat(years);
+
+    if (parseFloat(principal) <= 0) {
         alert("Enter a positive number");
-    }else{
-        document.getElementById("result").innerHTML = 
+    } else {
+        document.getElementById("result").innerHTML =
             `If you deposit <mark> ${principal}</mark>,<br>
             at an interest rate of  <mark> ${rate}%</mark><br>
             You will receive an amount of  <mark> ${amount}</mark>,<br>
@@ -20,12 +19,11 @@ function compute()
     document.getElementById("principal").focus();
 }
 
-function updateRate()
-{
-    document.getElementById("rateValue").innerText=document.getElementById("rate").value + "%";
-   // compute();
+function updateRate() {
+    document.getElementById("rateValue").innerText = document.getElementById("rate").value + "%";
+    // compute();
 }
 
-function fillTheBlanks(){
+function fillTheBlanks() {
     document.getElementById("principal").innerText = "100";
 }
